dagger-integration/requirements.txt
openai>=1.0.0
requests>=2.28.0
dagger-io>=0.9.0